pyproject.toml
[project]
name = "rubiksolver"
version = "0.1.0"
description = "GUI application to detect and solve a Rubik's cube"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kociemba>=1.2.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyopengl>=3.1.9",
    "pyrr>=0.10.3",
    "pyside6>=6.9.1",
    "superqt>=0.7.6",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.5",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "snakeviz>=2.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rubiksolver"]

[project.scripts]
rubiksolver = "rubiksolver.app:main"

[tool.pytest.ini_options]
addopts = "--maxfail=1 --cov-report html --cov"
